Refac: move auth mechanism to auth.py module
